+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added form input verification and corresponding test battery for site,state,and coordinate input
 - inactive location parameter forms are now hidden
 - added input form for hydrological unit codes
+- added data fetch script (npm run fetch)
 ### Changed